+++ CHANGELOG.md
@@ -1,8 +1,13 @@
+## 0.21.0
+- `QuestionEdit` uses a default value for the question parameter, able to be reused for `MQFEdit` and `MQFCreate`
+- TODO: Implement saving the new MQF
+- TODO: Implement the add question, add option, and delete question, and delete options for the `QuestionEdit` component
+
 ## 0.20.0
 - Used `React.useRef` for MDS and Name in `MQFEdit` and `MQFCreate`
 - Implemented the `TestDetails` component in `MQFEdit` and `MQFCreate`
 - Began work on reusing the `QuestionEdit` component in the `MQFCreate` page
-- TODO: Modify `QuestionEdit` so that it can be blank or take a question
+- ~~TODO: Modify `QuestionEdit` so that it can be blank or take a question~~
 
 ## 0.19.0
 - Moved Test Details from `/src/pages/MQFEdit` to own component for reusability